pyproject.toml
[tool.poetry]
name = "bondar"
version = "0.1.0"
description = ""
authors = ["Elijahrong <nhkim5760@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "0.17.0"
black = "21.12b0"
django-stubs = "1.9.0"
mypy = "0.931"
isort = "5.10.1"
channels = "^3.0.4"
channels-redis = "^3.3.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.isort]
profile = "black"